+++ packages/ui/CircularMarquee/CircularMarquee.ts
@@ -9,7 +9,7 @@ export interface CircularMarqueeProps extends BaseProps {
 
 /**
  * CircularMarquee class.
- * @link https://ui.studiometa.dev/-/components/CircularMarquee/
+ * @link https://ui.studiometa.dev/components/CircularMarquee/
  */
 export class CircularMarquee extends Base<CircularMarqueeProps> {
   /**